Dump of my enhanced version of the Bobber archetype. A sample project archetype included as well. Enhancements from the original version:
1. Can process any resource, not only m2 poms.
2. Velocity-based templates.
3. Conditional template execution, check the 'dependsOnVar' setting in the example project's archetype.xml descriptor.
4. Nicer user prompts for any variables hinting the default value.

Install the archetype first, check the cmd files for the syntax to generate a project. The scripts are very spartan and may fail for you (but trivial to fix), provided just to give an idea on how to use this stuff, and are not meant to be included.

TODO: first and the very most - rebundle the archetype plugin under maven plugins group, so we don't have to fully qualify the command, nor set it in the settings.xml via plugin alias.

git-svn-id: file:///home/projects/mule/scm@3482 bf997673-6b11-0410-b953-e057580c5b09