MULE-1692 - I've introduced an additional stream method in the connector interface.  This is necessary for reliable repeated streaming, since generic code in the streaming message adapter accesses this method directly and resource-management tends to be in transport-specific methods.
I realise that in the future we may want more sophisticated resource control (and all the rest of the things protocols enable), but even then we need to clean up on low-level stream shutdown, so I suspect this mechanism will always be used.  If that is the case (and I think it is), we may not need the "unsafe" getOutputStream, and I've tentatively marked it as deprecated.

git-svn-id: file:///home/projects/mule/scm@6301 bf997673-6b11-0410-b953-e057580c5b09