pom refactoring for building the site.
Javadoc is disabled for now as it seems to break the build process and site generation in general.

Notice that the maven-enforcer-plugin is in its own profile now. It is enabled by default so the behaviour did not change. Even though the plugin offers a skip option, the plugin is still invoked as part of each and every module which tends to fill up the log over time. So you can disable the enforcer plugin completely now by specifying -DskipEnforcer=true on the commandline.


git-svn-id: file:///home/projects/mule/scm@9214 bf997673-6b11-0410-b953-e057580c5b09