Rework the HttpResponse class to use OutputHandler. This has two nice benefits: * Its fast for every scenario * It is much more convenient for those writing transports like myself. Now I can just return an OutputHandler and not have to cache everything in memory.
MULE-3076: Catch the MuleRuntimeException if the mule-test-exclusions url is invalid and no file can be found.
This happens when you don't have a local mule-test-exclusions file, there is one in a JAR and you can't actually 
create a file for it.

When copying a messageadapter's properties, catch the UniqueIdNotSupportedException.

Add support for transforming InputStreams to OutputHandlers.

Add support for transforming HttpRquest messages to InputStreams.



git-svn-id: file:///home/projects/mule/scm@11186 bf997673-6b11-0410-b953-e057580c5b09