Move all the examples from using xfire to using CXF. The XFire connector will be on the forge (later today or tomorrow).
One major outstanding issue: there is a problem with CXF sending two replyto
messages. I'll have a message to the list shortly about this. Hopefully it is
a simple fix. Until this is fixed, the loan broker ESN sample and the ReplyToTestCase
will not work.

The other issue is more of a buld issue. The CXF Dynamic client (i.e. wsdl-cxf:...)
requires tools.jar on the classpath. Since this isn't a guaranteed assumption,
I've excluded the tests that depend on this. They all work however. Any thoughts on this?

Other stuff:
- Improved the CXF transport's handling of transport headers via a new interceptor
- CxfServiceComponent now correctly handles GET messages
- The Cxf dispatcher now mimics the XFire dispatcher's functionality
- Migrated various tests to use CXF instead of XFire 
 


git-svn-id: file:///home/projects/mule/scm@11351 bf997673-6b11-0410-b953-e057580c5b09