MULE-3383
Because we rewrite the endpoint address inside CXF when the WSDL is generated,
we need to store the endpoints inside the Connector for easy lookup by the
original endpoint name.  

Also, we now correctly handle the case where a VM endpoint is in the middle
between an HTTP endpoint and a CXF one. Turns out there is a MULE_ENDPOINT_PROPERTY
that we can use to find the originating URL of the request and use that 
inside the WSDL!



git-svn-id: file:///home/projects/mule/scm@11868 bf997673-6b11-0410-b953-e057580c5b09