Support efficient transformations when proxying web services. We can now take a StAX message, send it into SAXON as SAX events, then write the resulting SAX events directly to the web service response.
This magic is done via a new class called DelayedResult. Because
we don't want to have to write the transformed XML to memory,
we create a DelayedResult which allows someone down the stream
to control how the XML is written. In this case, CXF passes in
a new Result so SAXON can write directly to the XMLStreamWriter.

I had to change DefaultMuleMessage so it recognized other 
consumable types besides InputStream. Feedback on whether or not
there is a better way to do this is immensely welcome.


git-svn-id: file:///home/projects/mule/scm@12248 bf997673-6b11-0410-b953-e057580c5b09